+++ sw.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = 'learningenglish-v27';
+const CACHE_NAME = 'learningenglish-v28';
 const ASSETS_TO_CACHE = [
     './',
     './index.html',